LayoutTests/platform/efl/accessibility/legend-expected.txt

Choose a shipping method:
Overnight
End of test
This tests that a fieldset's title ui element is the legend.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


AXRole: AXWebArea 
    AXRole: AXGroup 
        AXRole: AXLabel AXValue: Choose a shipping method:     << fieldset's titleUIElement
        AXRole: AXRadioButton AXValue: 
    AXRole: AXDiv AXValue: End of test
PASS titleUIElement != null is true
PASS titleUIElementText.stringValue is 'AXValue: Choose a shipping method:'
PASS successfullyParsed is true

TEST COMPLETE
